Created new sturucture and files
