+++ server/app/api/v1/artifacts.py
@@ -20,9 +20,7 @@ Artifact API endpoints and business logic.
 This module contains artifact-related API endpoints and their business logic,
 including artifact types.
 """
-import os
-import json
-from fastapi import APIRouter, Depends, Request, HTTPException, UploadFile, File
+from fastapi import APIRouter, Depends, Request
 from sqlalchemy.ext.asyncio import AsyncSession
 from server.app.db.dbconfig import get_db
 from server.app.db.dbqueries import (
@@ -38,7 +36,6 @@ from server.app.services.mlmd_state import MlmdState
 from server.app.get_data import (
     get_artifact_types,
     async_api,
-    get_model_data,
 )
 import pandas as pd
 router = APIRouter(prefix="/v1", tags=["artifacts"])
@@ -46,7 +43,7 @@ router = APIRouter(prefix="/v1", tags=["artifacts"])
 # ==================== Business Logic Functions ====================
 
 # This API returns a list of artifact types in the current MLMD store.
-async def artifact_types(state: MlmdState):
+async def get_artifacts_types(state: MlmdState):
     """Get list of artifact types."""
     await state.check_mlmd_file_exists()
 
@@ -59,89 +56,6 @@ async def artifact_types(state: MlmdState):
         artifact_types_list.remove("Environment")
 
     return artifact_types_list
-
-# This API returns the model card payload including model, execution, and artifact data.
-async def model_card(state: MlmdState, modelId: int):
-    """Get model card information."""
-    json_payload_1 = ""
-    json_payload_2 = ""
-    json_payload_3 = ""
-    json_payload_4 = ""
-    model_data_df = pd.DataFrame()
-    model_exe_df = pd.DataFrame()
-    model_input_art_df = pd.DataFrame()
-    model_output_art_df = pd.DataFrame()
-    # checks if mlmd file exists on server
-    await state.check_mlmd_file_exists()
-    model_data_df, model_exe_df, model_input_art_df, model_output_art_df = await async_api(
-        get_model_data, state.query, modelId
-    )
-    if not model_data_df.empty:
-        result_1 = model_data_df.to_json(orient="records")
-        json_payload_1 = json.loads(result_1)
-    if not model_exe_df.empty:
-        result_2 = model_exe_df.to_json(orient="records")
-        json_payload_2 = json.loads(result_2)
-    if not model_input_art_df.empty:
-        result_3 = model_input_art_df.to_json(orient="records")
-        json_payload_3 = json.loads(result_3)
-    if not model_output_art_df.empty:
-        result_4 = model_output_art_df.to_json(orient="records")
-        json_payload_4 = json.loads(result_4)
-    return [json_payload_1, json_payload_2, json_payload_3, json_payload_4]
-
-
-async def upload_label(file: UploadFile):
-    """Upload label file."""
-    try:
-        if not file.filename:
-            raise HTTPException(status_code=400, detail="No file provided.")
-
-        labels_dir = "/cmf-server/data/labels"
-        file_path = os.path.join(labels_dir, os.path.basename(file.filename))
-
-        os.makedirs(os.path.dirname(file_path), exist_ok=True)
-
-        if os.path.exists(file_path):
-            return {
-                "message": f"File '{file.filename}' already exists at {labels_dir}. Skipping upload."
-            }
-
-        with open(file_path, "wb") as buffer:
-            buffer.write(await file.read())
-
-        return {"message": f"File '{file.filename}' uploaded successfully to {labels_dir}."}
-
-    except Exception as e:
-        return {"error": f"Failed to upload file: {e}"}
-
-
-"""Retrieve label data from from the /cmf-server/data/labels folder."""
-async def get_label_data(file_name: str) -> str:
-   
-    """
-    API endpoint to fetch the content of a requirements file.
-
-    Args:
-        file_name (str): The name of the file to be fetched. Must end with .csv.
-
-    Returns:
-        str: The content of the file as plain text.
-
-    Raises:
-        HTTPException: If the file does not exist or the extension is unsupported.
-    """
-    # Check if the file exists
-    file_path = os.path.join("/cmf-server/data/labels/", os.path.basename(file_name))
-    if not os.path.exists(file_path):
-        raise HTTPException(status_code=404, detail="File not found")
-    # Read and return the file content as plain text
-    try:
-        with open(file_path, "r") as file:
-            content = file.read()
-        return content
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Error reading file: {str(e)}")
 
 
 async def get_artifact_types_by_stage(
@@ -220,7 +134,7 @@ async def get_artifacts_by_stage(
 
 # ==================== API Endpoints ====================
 # only This API is used by the MCP server.
-@router.get("/artifacts/artifact-types")
+@router.get("/artifacts/artifact/types")
 async def get_artifacts_by_types(
     request: Request,
 ):
@@ -228,7 +142,7 @@ async def get_artifacts_by_types(
     Retrieve available artifact types.
     """
     state = request.app.state.mlmd
-    result = await artifact_types(state)
+    result = await get_artifacts_types(state)
 
     return success_response(
         data=result,
@@ -237,39 +151,7 @@ async def get_artifacts_by_types(
     )
 
 
-@router.get("/artifacts/model-card")
-async def get_model_card( request: Request, modelId: int):
-    state = request.app.state.mlmd
-    result = await model_card(state, modelId)
-    return success_response(
-        data=result,
-        message="Model card retrieved successfully",
-        code=200
-    )
-
-@router.post("/artifacts/label")
-async def upload_label_file(file: UploadFile = File(..., description="The file to upload")):
-    result = await upload_label(file)
-
-    return success_response(
-        data=result,
-        message="Label uploaded successfully",
-        code=201
-    )
-
-
-@router.get("/artifacts/label-data")
-async def get_label_data_route(file_name: str):
-    result = await get_label_data(file_name)
-
-    return success_response(
-        data=result,
-        message="Label data retrieved successfully",
-        code=200
-    )
-
-
-@router.post("/pipelines/{pipeline_name}/artifacts/types")
+@router.post("/pipelines/{pipeline_name}/artifacts/stages/{stage}/types")
 async def get_artifact_types_by_stage_route(
     query_params: ArtifactTypesByStageRequest,
     db: AsyncSession = Depends(get_db)
@@ -286,7 +168,7 @@ async def get_artifact_types_by_stage_route(
     )
 
 
-@router.post("/pipelines/{pipeline_name}/artifacts")
+@router.post("/pipelines/{pipeline_name}/artifacts/stages/{stage}")
 async def get_artifacts_by_stage_route(
     query_params: ArtifactByStageRequest,
     db: AsyncSession = Depends(get_db)
